moved mock patients to server

+++ server/src/index.ts
@@ -3,7 +3,9 @@ import express from "express";
 import { config } from "./config";
 import mongoose from "mongoose";
 
-import { UserModel } from "./models";
+import { UserModel } from "./user";
+import { PatientModel } from './patient';
+
 import { router } from "./routes";
 
 const app = express();
@@ -18,9 +20,17 @@ mongoose.connection.on('error', function(err: Error) {
  console.error('MongoDB connection error: ' + err);
 });
 
-// populate db, move this later to run only once
-UserModel.collection.deleteMany({}).then(() => console.log("All users deleted"));
-UserModel.collection.insertMany(users).then(() => console.log("Inserted users from JSON"));
+// populate db with users and patients
+mongoose.connection.on('open', function() {
+    mongoose.connection.db.listCollections({name: 'UserModel'})
+        .next(function(err, collinfo) {
+            if (collinfo) {
+                UserModel.collection.deleteMany({}).then(() => console.log("All users deleted"));
+                UserModel.collection.insertMany(users).then(() => console.log("Inserted users from JSON"));
+            }
+        });
+});
+
  
 // Cross Origin middleware
 app.use(function(req: express.Request, res: express.Response, next: express.NextFunction) {
